requirements.txt
flask>=2.3.0
pdfplumber>=0.10.0
python-docx>=0.8.11
PyPDF2>=3.0.0
